+++ CHANGELOG.markdown
@@ -3,6 +3,10 @@
 
 ## Version 0.9.1 - *Not Yet Released*
 
+ * Added Aphid.Model#identifierAttribute for specifying the name of the
+   attribute that should be considered the identifier for the model to be used
+   when loading, reloading and saving.
+
  * Added a custom String#toDate extension to be used for converting dates
    formatted as strings to instances of Date.